Cleanup, and a few small fixes.
* Fixed qscan display error on expert mode #1405
* Fixed issue where setting a date < 1970 can crash the bbs
* Added tests in DateTime for invalid dates at construction
* Removed unneeded header in status.h, added them into the handful of files
  that needed that include.

Fixes #1405

+++ bbs/showfiles.cpp
@@ -55,7 +55,7 @@ void show_files(const char *file_name, const char *pszDirectoryName) {
   FindFiles ff(full_pathname, FindFiles::FindFilesType::files);
   for (const auto& f : ff) {
     full_pathname = StrCat("|#7[|#2", aligns(f.name), "|#7]|#1 ");
-    if (bout.wherex() > static_cast<int>(a()->user()->GetScreenChars() - 15)) {
+    if (bout.wherex() > a()->user()->GetScreenChars() - 15) {
       bout.nl();
     }
     bout << full_pathname;
